requirements/base.txt
Django==4.1.2
psycopg2-binary==2.9.4
django-environ==0.9.0